Refactor internal/os into x/wasm
